# backend/docs/swagger.yaml
basePath: /
definitions:
...
  CreateCategoryRes:
    properties:
      data:
        $ref: '#/definitions/domain.Category'
      total_amount:
        format: double
        type: number
      transaction_count:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/domain.Transaction'
        type: array
    required:
    - data
    type: object
...
  DeleteCategoryRes:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  UpdateCategoryRes:
    properties:
      data:
        $ref: '#/definitions/domain.Category'
      total_amount:
        format: double
        type: number
      transaction_count:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/domain.Transaction'
        type: array
    required:
    - data
    type: object
...
  category.CreateCategoryReq:
    properties:
      name:
        type: string
      type:
        type: string
    required:
    - name
    - type
    type: object
  category.ListCategoriesRes:
    properties:
      categories:
        items:
          $ref: '#/definitions/dto.Category'
        type: array
    required:
    - categories
    type: object
  category.UpdateCategoryReq:
    properties:
      name:
        type: string
      type:
        type: string
    type: object
  domain.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        description: income, expense
        type: string
      updated_at:
        type: string
      user_id:
        description: nil for default categories
        type: integer
    required:
    - created_at
    - id
    - name
    - type
    - updated_at
    type: object
...
  domain.Transaction:
    properties:
      amount:
        format: double
        type: number
      category_id:
        type: integer
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      wallet_id:
        type: integer
    required:
    - amount
    - category_id
    - created_at
    - date
    - description
    - id
    - updated_at
    - wallet_id
    type: object
...
  dto.Category:
    properties:
      data:
        $ref: '#/definitions/domain.Category'
      total_amount:
        format: double
        type: number
      transaction_count:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/domain.Transaction'
        type: array
    required:
    - data
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API untuk aplikasi manajemen keuangan dan pencatatan emas - ARTA
  title: ARTA API
  version: "1.0"
paths:
...
  /api/category:
    get:
      consumes:
      - application/json
      operationId: ListCategories
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: income or expense
        in: query
        name: type
        type: string
      - description: wallet_id
        in: query
        name: wallet_id
        type: integer
      - description: include transaction stats
        in: query
        name: include_stats
        type: boolean
      - description: start_date (2006-01-02)
        in: query
        name: start_date
        type: string
      - description: end_date (2006-01-02)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.ListCategoriesRes'
      tags:
      - category
    post:
      consumes:
      - application/json
      operationId: CreateCategory
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Unique key per submission attempt for safe retry (UUID recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/category.CreateCategoryReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreateCategoryRes'
      tags:
      - category
  /api/category/{category_id}:
    get:
      consumes:
      - application/json
      operationId: GetCategory
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: category id
        in: path
        name: category_id
        required: true
        type: integer
      - description: wallet_id
        in: query
        name: wallet_id
        type: integer
      - description: include_total_amount
        in: query
        name: include_total_amount
        type: boolean
      - description: include_transactions
        in: query
        name: include_transactions
        type: boolean
      - description: start_date
        in: query
        name: start_date
        type: string
      - description: end_date
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Category'
      tags:
      - category
  /api/category/{id}:
    delete:
      consumes:
      - application/json
      operationId: DeleteCategory
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: category id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DeleteCategoryRes'
      tags:
      - category
    put:
      consumes:
      - application/json
      operationId: UpdateCategory
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: category id
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/category.UpdateCategoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UpdateCategoryRes'
      tags:
      - category
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"